Fix math logic error in alogg_poll_ogg_ts() which might be the cause of static during playback.

+++ trunk/src/alogg/src/alogg.c
@@ -506,13 +506,13 @@ int alogg_poll_ogg_ts(ALOGG_OGG *ogg) {
     /* process samples with Rubber Band */
     if (ogg->stereo) {
       for (i = 0; i < ogg->time_stretch_buffer_samples; i++) {
-        ogg->time_stretch_buffer[0][i] = (float)(audiobuf_sp[i * 2] - 0x8000) / (float)0x8000;
-        ogg->time_stretch_buffer[1][i] = (float)(audiobuf_sp[i * 2 + 1] - 0x8000) / (float)0x8000;
+        ogg->time_stretch_buffer[0][i] = (float)((long)audiobuf_sp[i * 2] - 0x8000) / (float)0x8000;
+        ogg->time_stretch_buffer[1][i] = (float)((long)audiobuf_sp[i * 2 + 1] - 0x8000) / (float)0x8000;
       }
     }
     else {
       for (i = 0; i < ogg->time_stretch_buffer_samples; i++) {
-        ogg->time_stretch_buffer[0][i] = (float)(audiobuf_sp[i] - 0x8000) / (float)0x8000;
+        ogg->time_stretch_buffer[0][i] = (float)((long)audiobuf_sp[i] - 0x8000) / (float)0x8000;
       }
 	}
     rubberband_process(ogg->time_stretch_state, (const float **)ogg->time_stretch_buffer, ogg->time_stretch_buffer_samples, 0);
@@ -522,13 +522,13 @@ int alogg_poll_ogg_ts(ALOGG_OGG *ogg) {
   size_done = rubberband_retrieve(ogg->time_stretch_state, ogg->time_stretch_buffer, ogg->time_stretch_buffer_samples);
   if (ogg->stereo) {
     for (i = 0; i < size_done; i++) {
-      audiobuf_sp[i * 2] = ogg->time_stretch_buffer[0][i] * (float)0x8000 + 0x8000;
-      audiobuf_sp[i * 2 + 1] = ogg->time_stretch_buffer[1][i] * (float)0x8000 + 0x8000;
+      audiobuf_sp[i * 2] = (float)ogg->time_stretch_buffer[0][i] * (float)0x8000 + (float)0x8000;
+      audiobuf_sp[i * 2 + 1] = (float)ogg->time_stretch_buffer[1][i] * (float)0x8000 + (float)0x8000;
     }
   }
   else {
     for (i = 0; i < size_done; i++) {
-      audiobuf_sp[i] = ogg->time_stretch_buffer[0][i] * (float)0x8000 + 0x8000;
+      audiobuf_sp[i] = (float)ogg->time_stretch_buffer[0][i] * (float)0x8000 + (float)0x8000;
     }
   }
